refactor: change the return type of the OSXSys::get_change_count to isize and allow dead code (#14)
Since the OSXSys::get_change_count never failes, the return type dose not need to be Result.
Furthermore, the Observer dose not need to send an Error to the ClipboardStream(receiver).
BodySenders::send_all_if_some is no longer used due to this changes. However, since this method will still be used after the refactoring, we've left it as is.